+++ plugins/mapper/src/org/ietr/preesm/plugin/mapper/model/impl/TransferVertex.java
@@ -48,7 +48,7 @@ import org.sdf4j.model.dag.DAGEdge;
  */
 public class TransferVertex extends MapperDAGVertex {
 
-	public static final int SEND_RECEIVE_COST = 100;
+	public static final long SEND_RECEIVE_COST = 100;
 	
 	private RouteStep step;
