+++ plugins/mapper/src/org/ietr/preesm/plugin/mapper/plot/GanttPlotter.java
@@ -374,7 +374,7 @@ public class GanttPlotter extends ApplicationFrame {
 
 		logger.log(Level.FINEST, "Getting finishing times");
 
-		int test;
+		long test;
 
 
 		simulator.retrieveTotalOrder();
