+++ plugins/mapper/src/org/ietr/preesm/plugin/mapper/test/ListSchedulingArchiTester.java
@@ -162,7 +162,7 @@ public class ListSchedulingArchiTester {
 		PreesmLogger.getLogger().log(Level.FINEST, "Displaying dag implanted ");
 		scheduler.dagimplanteddisplay(dag, simu2);
 
-		int finalTime = simu2.getFinalTime();
+		long finalTime = simu2.getFinalTime();
 		PreesmLogger.getLogger().log(
 				Level.FINER,
 				"DAG Final Time on " + archi.getNumberOfOperators()
